# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentOut": {
        "properties": {
          "can_delete": {
            "title": "Can Delete",
            "type": "boolean"
          },
          "comment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment Id"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "is_image": {
            "title": "Is Image",
            "type": "boolean"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "uploader": {
            "$ref": "#/components/schemas/UserPublic"
          }
        },
        "required": [
          "id",
          "incident_id",
          "comment_id",
          "filename",
          "content_type",
          "size_bytes",
          "is_image",
          "uploader",
          "created_at",
          "can_delete"
        ],
        "title": "AttachmentOut",
        "type": "object"
      },
...
      "DownloadLink": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_in"
        ],
        "title": "DownloadLink",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserPublic": {
        "description": "What any signed-in user may see about a teammate (avatars, names, mentions).",
        "properties": {
          "avatar_color": {
            "title": "Avatar Color",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "avatar_color"
        ],
        "title": "UserPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-assisted incident management. All business routes live under /api/v1.",
    "title": "Tasky API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/attachments/{attachment_id}": {
      "delete": {
        "operationId": "delete_api_v1_attachments__attachment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attachment Id",
              "type": "string"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete an attachment (uploader or admin)",
        "tags": [
          "attachments"
        ]
      }
    },
    "/api/v1/attachments/{attachment_id}/download": {
      "get": {
        "description": "Redirects to a presigned URL. Send `Accept: application/json` to get the link as JSON instead (for <img> previews, which cannot send a bearer token).",
        "operationId": "download_api_v1_attachments__attachment_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attachment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inline",
            "required": false,
            "schema": {
              "default": false,
              "title": "Inline",
              "type": "boolean"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadLink"
                }
              }
            },
            "description": "Successful Response"
          },
          "302": {
            "description": "Redirect to the file"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download (302 to a short-lived link)",
        "tags": [
          "attachments"
        ]
      }
    },
...
    "/api/v1/incidents/{ident}/attachments": {
      "get": {
        "operationId": "list_attachments_api_v1_incidents__ident__attachments_get",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttachmentOut"
                  },
                  "title": "Response List Attachments Api V1 Incidents  Ident  Attachments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List attachments",
        "tags": [
          "attachments"
        ]
      },
      "post": {
        "description": "Body: the raw file. Allowed: PNG, JPEG, GIF, PDF, TXT, LOG, JSON; at most MAX_UPLOAD_MB. The type is detected from the content, not the header.",
        "operationId": "upload_api_v1_incidents__ident__attachments_post",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "comment_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Comment Id"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload an attachment",
        "tags": [
          "attachments"
        ]
      }
    },
...
  }
}